requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
python-dotenv==1.0.1
faker==22.6.0  # For generating dummy data 